create both pipeline from same function
