+++ pipeline/src/indices/events.ts
@@ -1,1 +1,5 @@
 export const index = "events";
+
+export const mapping = {};
+
+export const settings = {};
